+++ cmd/key_generate.go
@@ -176,5 +176,5 @@ func PrintGenerateKeyResponse(w io.Writer, resp *wallet.GenerateKeyResponse) {
 
 	p.BlueArrow().InfoText("Run the service").NextLine()
 	p.Text("Now, you can run the service. See the following command:").NextSection()
-	p.Code(fmt.Sprintf("%s service run --help", os.Args[0])).NextSection()
+	p.Code(fmt.Sprintf("%s service run --help", os.Args[0])).NextLine()
 }