+++ cmd/key_taint.go
@@ -130,5 +130,5 @@ func PrintTaintKeyResponse(w io.Writer) {
 
 	p.BlueArrow().InfoText("Untaint a key").NextLine()
 	p.Text("You may have tainted a key pair by mistake. If you want to untaint it, see the following command:").NextSection()
-	p.Code(fmt.Sprintf("%s key untaint --help", os.Args[0])).NextSection()
+	p.Code(fmt.Sprintf("%s key untaint --help", os.Args[0])).NextLine()
 }