+++ cmd/wallet_create.go
@@ -130,5 +130,5 @@ func PrintCreateWalletResponse(w io.Writer, resp *wallet.CreateWalletResponse) {
 
 	p.BlueArrow().InfoText("Run the service").NextLine()
 	p.Text("Now, you can run the service. See the following command:").NextSection()
-	p.Code(fmt.Sprintf("%s service run --help", os.Args[0])).NextSection()
+	p.Code(fmt.Sprintf("%s service run --help", os.Args[0])).NextLine()
 }